src/test/java/com/leetcode/strings/easy/RansomNote_383Test.java
package com.leetcode.strings.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class RansomNote_383Test {
  private final RansomNote_383 solution = new RansomNote_383();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments("a", "b", false), arguments("aa", "ab", false), arguments("aa", "aab", true));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void canConstruct(String note, String magazine, boolean expectedResult) {
    boolean actualResult = solution.canConstruct(note, magazine);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
